+++ src/app/Fake.BuildServer.TeamCity/TeamCity.fs
@@ -257,6 +257,8 @@ module TeamCity =
                     testFailed testName message detail
                 | TraceData.TestStatus (testName,TestStatus.Failed(message, detail, Some (expected, actual))) ->
                     comparisonFailure testName message detail expected actual
+                | TraceData.BuildState state ->
+                    ConsoleWriter.write false color true (sprintf "Changing BuildState to: %A" state)
                 | TraceData.CloseTag (KnownTags.Test name, time, _) ->
                     finishTestCase name time
                 | TraceData.OpenTag (KnownTags.TestSuite name, _) ->